Update test scripts to work with GH actions
For `test`, just run PHPUnit, which will use the extension if it's present (so
don't use it when you expect the extension to be present).

For `test-no-ini`, specify the `dom` extension, which PHPUnit requires. You can
use this script even when the extension is installed because we don't tell PHP
to load it.