Improve Psalm typing
Trying to use the polyfill with a codebase that uses Psalm resulted in a bunch
of errors due to the Psalm typing on these classes. We should be using the same
types that the Psalm stub uses for the extension.

This doesn't do that yet. This was just a start on fixing the type errors that
our code base happened to reveal. Applying the types from the stub would be
much better.